+++ sdk/resourcemanager/signalr/armsignalr/privatelinkresources_client_example_test.go
@@ -24,11 +24,11 @@ func ExamplePrivateLinkResourcesClient_NewListPager() {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client, err := armsignalr.NewPrivateLinkResourcesClient("00000000-0000-0000-0000-000000000000", cred, nil)
+	clientFactory, err := armsignalr.NewClientFactory("<subscription-id>", cred, nil)
 	if err != nil {
 		log.Fatalf("failed to create client: %v", err)
 	}
-	pager := client.NewListPager("myResourceGroup", "mySignalRService", nil)
+	pager := clientFactory.NewPrivateLinkResourcesClient().NewListPager("myResourceGroup", "mySignalRService", nil)
 	for pager.More() {
 		page, err := pager.NextPage(ctx)
 		if err != nil {