+++ sdk/resourcemanager/signalr/armsignalr/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 (2023-03-24)
 ### Features Added
 
+- New struct `ClientFactory` which is a client factory used to create any client in this module
 - New struct `ServerlessSettings`
 - New field `Serverless` in struct `Properties`